requirements.txt
annotated-types
anyio
click
fastapi
h11
idna
pydantic
pydantic_core
sniffio
starlette
typing_extensions
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic